requirements/base.txt
# Python Version 3.11
# ----------------------------------------------
Django==4.2.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-filter==23.2
django-cors-headers==4.0.0
django-environ==0.10.0
drf-yasg==1.21.5
pyotp==2.8.0
